requirements.txt
requests==<latest_version>
python-dotenv==<latest_version>
